Splitting commands, programs, and utils into individual files. This should make it easier to modify particular features and isolate the files being modified.
