+++ mycommands/command/listall.py
@@ -1,85 +1,5 @@
-"""Command module.
-Useful commands for myshell to manage the presentation of information in the terminal, and to find information on programs and classes to which myshell has access."""
-
-import re
-import subprocess
-from os import name as os_name
-
-from myprograms.myprograms import MyProgram
-
-
-
-class MyCommand:
-    """Command object."""
-    @classmethod    
-    def help(command):
-        """Prints the `__doc__` string of the class.
-        
-        First cleans out any backtick marks, and `ReST`-type dockstring lines."""
-        try:
-            help = re.sub("`", "", command.__doc__)
-            # ":param int number:" becomes "number:"
-            help = re.sub(r":param \w* ", "", help)
-            print(help)
-        except Exception as e:
-            raise Exception(f"An unexpected error occurred with printing help info: {e}")
-
-
-
-class ClearTerminal(MyCommand):
-    """Clears the terminal. Works on posix and nt-like systems.
-    
-    :param str name: `\"clear\"`
-    :param lst aliases: `[\"cls\"]`"""
-    name = "Clear"
-    aliases = ["cls"]
-
-    def __init__(self):
-        """creates a ClearTerminal object.
-
-        :param classmethod operation: set to `Clearterminal.ClearTerminal`
-        """
-        self.operation = self.ClearTerminal
-
-    def getos_name(self):
-        """Determines what operating system namespace is ClearTerminal being run on."""
-        return os_name
-
-    def run(self, arg, shell=False):
-        """Allows processes to run commands in terminal with arguments."""
-        subprocess.run(arg,shell=shell)
-
-    # While "java" is registered in os.name, jython does not yet support python 3. So I'm not bothering with that.
-    def ClearTerminal(self):
-        """For nt and posix systems, attempts to run a clear terminal command through a subprocess."""
-        try:
-            if self.getos_name() == "nt":
-                self.run('cls', shell = True)
-            else:
-                self.run("clear")
-        except Exception as e:
-            raise Exception(f"An unexpected error occurred with ClearTerminal: {e}")
-
-
-
-class PageBreak(MyCommand):
-    """Prints a long line and then moves to the next line of the terminal.
-    
-    :param str name: `"PageBreak"`"""
-
-    name = "PageBreak"
-
-    def __init__(self):
-        """Creates a PageBreak object.
-        
-        :param classmethod operation: points to `self.PageBreak`."""
-        self.operation = self.PageBreak
-
-    def PageBreak(self):
-        try:
-            print("—————————————————————————————")
-        except Exception as e:
-            raise Exception(f"An unexpected error occurred with \"PageBreak\": {e}")
+from mycommands import MyCommand
+from myprograms import MyProgram
