+++ myprograms/__init__.py
@@ -1,1 +1,13 @@
-from .myprograms import *
+"""Program module.
+Programs for the user to run, perhaps with one or more instances of input."""
+
+
+from myprograms.core import MyProgram
+
+from .program.binarymasktodecimal import BinaryMaskToDecimalProgram
+from .program.decimaltobinarymask import DecimalToBinaryMaskProgram
+
+
+
+
+
